move a few minigames to slash commands
