--- src/commands/Minion/castlewars.ts
@@ -1,53 +1,0 @@
-import { Time } from 'e';
-import { CommandStore, KlasaMessage } from 'klasa';
-
-import { minionNotBusy, requiresMinion } from '../../lib/minions/decorators';
-import { getMinigameEntity } from '../../lib/settings/settings';
-import { BotCommand } from '../../lib/structures/BotCommand';
-import { MinigameActivityTaskOptions } from '../../lib/types/minions';
-import { formatDuration } from '../../lib/util';
-import addSubTaskToActivityTask from '../../lib/util/addSubTaskToActivityTask';
-
-export default class CastleWarsCommand extends BotCommand {
-	public constructor(store: CommandStore, file: string[], directory: string) {
-		super(store, file, directory, {
-			altProtection: true,
-			description: 'Sends your minion to play Castle Wars.',
-			usage: '[play] [name:str]',
-			examples: ['+cw'],
-			categoryFlags: ['minion', 'minigame'],
-			subcommands: true,
-			aliases: ['cw']
-		});
-	}
-
-	async run(msg: KlasaMessage) {
-		const bank = msg.author.bank();
-		const kc = await getMinigameEntity(msg.author.id);
-		return msg.channel.send(`You have **${bank.amount('Castle wars ticket')}** Castle wars tickets.
-You have played ${kc.castle_wars} Castle Wars games.`);
-	}
-
-	@requiresMinion
-	@minionNotBusy
-	async play(msg: KlasaMessage) {
-		const gameLength = Time.Minute * 18;
-		const quantity = Math.floor(msg.author.maxTripLength('CastleWars') / gameLength);
-		const duration = quantity * gameLength;
-
-		await addSubTaskToActivityTask<MinigameActivityTaskOptions>({
-			userID: msg.author.id,
-			channelID: msg.channel.id,
-			duration,
-			type: 'CastleWars',
-			quantity,
-			minigameID: 'castle_wars'
-		});
-
-		return msg.channel.send(
-			`${
-				msg.author.minionName
-			} is now doing ${quantity} games of Castle Wars. The trip will take around ${formatDuration(duration)}.`
-		);
-	}
-}
