+++ src/mahoji/commands/minigames.ts
@@ -1,6 +1,7 @@
 import { ApplicationCommandOptionType, CommandRunOptions } from 'mahoji';
 
 import { client } from '../..';
+import { LMSBuyables } from '../../lib/data/CollectionsExport';
 import {
 	barbAssaultGambleCommand,
 	barbAssaultLevelCommand,
@@ -8,6 +9,15 @@ import {
 	barbAssaultStatsCommand,
 	GambleTiers
 } from '../lib/abstracted_commands/barbAssault';
+import { castleWarsStartCommand, castleWarsStatsCommand } from '../lib/abstracted_commands/castleWarsCommand';
+import { lmsCommand } from '../lib/abstracted_commands/lmsCommand';
+import {
+	pestControlBuyables,
+	pestControlBuyCommand,
+	pestControlStartCommand,
+	pestControlStatsCommand,
+	pestControlXPCommand
+} from '../lib/abstracted_commands/pestControlCommand';
 import { OSBMahojiCommand } from '../lib/util';
 import { mahojiUsersSettingsFetch } from '../mahojiSettings';
 
@@ -15,10 +25,15 @@ export const minigamesCommand: OSBMahojiCommand = {
 	name: 'minigames',
 	description: 'Send your minion to do various minigames.',
 	options: [
+		/**
+		 *
+		 * Barbarian Assault
+		 *
+		 */
 		{
 			type: ApplicationCommandOptionType.SubcommandGroup,
 			name: 'barb_assault',
-			description: 'The Barbarian Assault minigame. (Solo)',
+			description: 'The Barbarian Assault minigame.',
 			options: [
 				{
 					type: ApplicationCommandOptionType.Subcommand,
@@ -57,6 +72,153 @@ export const minigamesCommand: OSBMahojiCommand = {
 					description: 'See your Barbarian Assault stats.'
 				}
 			]
+		},
+		/**
+		 *
+		 * Castle Wars
+		 *
+		 */
+		{
+			type: ApplicationCommandOptionType.SubcommandGroup,
+			name: 'castle_wars',
+			description: 'The Castle Wars minigame.',
+			options: [
+				{
+					type: ApplicationCommandOptionType.Subcommand,
+					name: 'stats',
+					description: 'See your Castle Wars stats.'
+				},
+				{
+					type: ApplicationCommandOptionType.Subcommand,
+					name: 'start',
+					description: 'Start a Castle Wars trip.'
+				}
+			]
+		},
+		/**
+		 *
+		 * LMS
+		 *
+		 */
+		{
+			name: 'lms',
+			description: 'Sends your minion to do the Last Man Standing minigame.',
+			type: ApplicationCommandOptionType.SubcommandGroup,
+			options: [
+				{
+					type: ApplicationCommandOptionType.Subcommand,
+					name: 'stats',
+					description: 'See your Last Man Standing stats.'
+				},
+				{
+					type: ApplicationCommandOptionType.Subcommand,
+					name: 'start',
+					description: 'Start a Last Man Standing Trip'
+				},
+				{
+					type: ApplicationCommandOptionType.Subcommand,
+					name: 'buy',
+					description: 'Buy a reward using your points.',
+					options: [
+						{
+							name: 'name',
+							description: 'The item you want to purchase.',
+							type: ApplicationCommandOptionType.String,
+							required: true,
+							autocomplete: async (value: string) => {
+								return LMSBuyables.filter(i =>
+									!value ? true : i.item.name.toLowerCase().includes(value.toLowerCase())
+								).map(i => ({ name: i.item.name, value: i.item.name }));
+							}
+						},
+						{
+							name: 'quantity',
+							description: 'The quantity you want to purchase.',
+							type: ApplicationCommandOptionType.Integer,
+							required: false,
+							min_value: 1,
+							max_value: 1000
+						}
+					]
+				},
+				{
+					type: ApplicationCommandOptionType.Subcommand,
+					name: 'simulate',
+					description: 'Simulate a Last Man Standing game with your Discord friends.',
+					options: [
+						{
+							name: 'names',
+							description: 'Names. e.g. Magnaboy, Kyra, Alex',
+							required: false,
+							type: ApplicationCommandOptionType.String
+						}
+					]
+				}
+			]
+		},
+		/**
+		 *
+		 * Pest Control
+		 *
+		 */ {
+			name: 'pest_control',
+			description: 'Sends your minion to do the Pest Control minigame.',
+			type: ApplicationCommandOptionType.SubcommandGroup,
+			options: [
+				{
+					type: ApplicationCommandOptionType.Subcommand,
+					name: 'stats',
+					description: 'See your Pest Control stats.'
+				},
+				{
+					type: ApplicationCommandOptionType.Subcommand,
+					name: 'start',
+					description: 'Send your minion to do Pest Control games.'
+				},
+				{
+					type: ApplicationCommandOptionType.Subcommand,
+					name: 'xp',
+					description: 'Buy XP with your Pest Control commendation games.',
+					options: [
+						{
+							type: ApplicationCommandOptionType.String,
+							name: 'skill',
+							required: true,
+							description: 'The skill you want XP in.',
+							choices: ['attack', 'strength ', 'defence', 'hitpoints', 'ranged', 'magic', 'prayer'].map(
+								i => ({ name: i, value: i })
+							)
+						},
+						{
+							type: ApplicationCommandOptionType.Integer,
+							name: 'amount',
+							description: 'The amount of points you want to spend.',
+							min_value: 1,
+							max_value: 1000
+						}
+					]
+				},
+				{
+					type: ApplicationCommandOptionType.Subcommand,
+					name: 'buy',
+					description: 'Buy items with your Pest Control commendation games.',
+					options: [
+						{
+							type: ApplicationCommandOptionType.String,
+							name: 'name',
+							required: true,
+							description: 'The skill you want XP in.',
+							autocomplete: async value => {
+								return pestControlBuyables
+									.filter(i =>
+										!value ? true : i.item.name.toLowerCase().includes(value.toLowerCase())
+									)
+									.map(i => ({ name: i.item.name, value: i.item.name }));
+							}
+						}
+					]
+				}
+			]
 		}
 	],
 	run: async ({
@@ -71,9 +233,28 @@ export const minigamesCommand: OSBMahojiCommand = {
 			gamble?: { tier: string; quantity: number };
 			stats?: {};
 		};
+		castle_wars?: { stats?: {}; start?: {} };
+		lms?: {
+			stats?: {};
+			start?: {};
+			buy?: { name?: string; quantity?: number };
+			simulate?: { names?: string };
+		};
+		pest_control?: {
+			stats?: {};
+			xp?: { skill: string; amount: number };
+			start?: {};
+			buy?: { name: string };
+		};
 	}>) => {
 		const klasaUser = await client.fetchUser(userID);
 		const user = await mahojiUsersSettingsFetch(userID);
+
+		/**
+		 *
+		 * Barbarian Assault
+		 *
+		 */
 		if (options.barb_assault?.start) {
 			return barbAssaultStartCommand(channelID, user, klasaUser);
 		}
@@ -93,6 +274,46 @@ export const minigamesCommand: OSBMahojiCommand = {
 			return barbAssaultStatsCommand(user);
 		}
 
+		/**
+		 *
+		 * Castle Wars
+		 *
+		 */
+		if (options.castle_wars?.stats) {
+			return castleWarsStatsCommand(klasaUser);
+		}
+		if (options.castle_wars?.start) {
+			return castleWarsStartCommand(klasaUser, channelID);
+		}
+
+		/**
+		 *
+		 * LMS
+		 *
+		 */
+		if (options.lms) return lmsCommand(options.lms, klasaUser, channelID, interaction);
+
+		/**
+		 *
+		 * LMS
+		 *
+		 */
+		if (options.pest_control?.stats) return pestControlStatsCommand(user);
+		if (options.pest_control?.xp) {
+			return pestControlXPCommand(
+				interaction,
+				klasaUser,
+				user,
+				options.pest_control.xp.skill,
+				options.pest_control.xp.amount
+			);
+		}
+		if (options.pest_control?.start) {
+			return pestControlStartCommand(klasaUser, channelID);
+		}
+		if (options.pest_control?.buy) {
+			return pestControlBuyCommand(klasaUser, user, options.pest_control.buy.name);
+		}
 		return 'Invalid command.';
 	}
 };
